refactor: rename ht_ft and file; new print ft
